fix: pulled out errors to error file --- The following errors was pulled out into `parser/src/error.rs`: - EvalError - IncompatibleMatrixSizes
